Fixed significant typo in Matrix::fromArray leading to unpredictable behaviour
`Matrix::fromArray` is supposed to map a linear array onto a Matrix by
"linearly filling" the Matrix from left to right and top to bottom. We
call this linear indexing of the matrix. However, instead of the map
from source to matrix being defined as `source(row * columnCount +
column) -> matrix(row, column)`, it was defined as `source(row * column
+ column) -> matrix(row, column)`, leading to incorrect and deleterious
behaviour in `fromArray`.